+++ src/geometry/vector/mod.rs
@@ -3,27 +3,27 @@ mod mutation;
 mod scalar_ops;
 
 #[derive(Debug, Clone, Copy, Default, PartialEq, PartialOrd)]
-pub struct Vector(f32, f32, f32);
+pub struct Vector(f64, f64, f64);
 
 pub type Color = Vector;
 pub type Position = Vector;
 pub type Direction = Vector;
 
 impl Vector {
-    pub fn new(x: f32, y: f32, z: f32) -> Self { Self(x, y, z) }
+    pub fn new(x: f64, y: f64, z: f64) -> Self { Self(x, y, z) }
 
-    pub fn x(&self) -> f32 { self.0 }
+    pub fn x(&self) -> f64 { self.0 }
 
-    pub fn y(&self) -> f32 { self.1 }
+    pub fn y(&self) -> f64 { self.1 }
 
-    pub fn z(&self) -> f32 { self.2 }
+    pub fn z(&self) -> f64 { self.2 }
 
-    pub fn dot(&self, other: Self) -> f32 {
+    pub fn dot(&self, other: Self) -> f64 {
         let factor = *self * other;
         factor.x() + factor.y() + factor.z()
     }
 
-    pub fn length(&self) -> f32 { self.dot(*self).sqrt() }
+    pub fn length(&self) -> f64 { self.dot(*self).sqrt() }
 
     pub fn unit(self) -> Self { self / self.length() }
 }
